requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
django-regex-field==2.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.3.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pipenv==2021.5.29
platformdirs==2.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.9.0
virtualenv-clone==0.5.7
whitenoise==5.3.0
